move fastsim addOn tests to use condDBv2. Add AutoCondGlobalTag version of HLT config for condDBv2
