+++ FastSimulation/Configuration/test/IntegrationTestFake_cfg.py
@@ -30,10 +30,10 @@ process.load("Configuration.Generator.H200ZZ4L_cfi")
 
 
 # Famos sequences (MC conditions, not Fake anymore!)
-process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_cff")
+process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_condDBv2_cff")
 process.load('FastSimulation.Configuration.Geometries_cff')
 #process.load("FastSimulation.Configuration.CommonInputs_cff")
-from Configuration.AlCa.autoCond import autoCond
+from Configuration.AlCa.autoCond_condDBv2 import autoCond
 process.GlobalTag.globaltag = autoCond['mc']
 # Allow reading of the tracker geometry from the DB
 #process.load('CalibTracker/Configuration/Tracker_DependentRecords_forGlobalTag_nofakes_cff')
