+++ FastSimulation/Configuration/test/IntegrationTest_cfg.py
@@ -18,7 +18,7 @@ process.load("Configuration.Generator.TTbar_cfi")
 
 # Famos sequences (NO HLT)
 #process.load("FastSimulation.Configuration.CommonInputs_cff")
-process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_cff")
+process.load("Configuration.StandardSequences.FrontierConditions_GlobalTag_condDBv2_cff")
 process.load('FastSimulation.Configuration.Geometries_cff')
 process.load("FastSimulation.Configuration.FamosSequences_cff")
 
@@ -35,7 +35,7 @@ process.famosSimHits.SimulateCalorimetry = True
 process.famosSimHits.SimulateTracking = True
 
 # Get frontier conditions    - not applied in the HCAL, see below
-from HLTrigger.Configuration.AutoCondGlobalTag import AutoCondGlobalTag
+from HLTrigger.Configuration.AutoCondGlobalTag_condDBv2 import AutoCondGlobalTag
 process.GlobalTag = AutoCondGlobalTag(process.GlobalTag,'auto:run1_mc')
 # Allow reading of the tracker geometry from the DB
 process.load('CalibTracker/Configuration/Tracker_DependentRecords_forGlobalTag_nofakes_cff')
